Introduce tripleo_ceph_set_container_cli role
This role sets a new ceph_cli fact getting rid
of the cephadm binary. In fact, it's now possible
to run a octopus container which acts as a ceph
client.
By using a role it's possible to customize the
container options (e.g., changing fsid or ceph.conf
file it can interact with different ceph clusters).
The site.yaml is updated with this new role and the
cephadm rpm package is fixed as well.

Signed-off-by: Francesco Pantano <fpantano@redhat.com>