Improve Ceph services spec definition
The purpose of this commit is to improve the way the spec
and the hostmap are built.
Now there are three models (according to [1]):

1. by pattern
2. by label
3. by hosts

Sounds like OSDs support by pattern only, hence the spec
is aligned with this model.
Further, for each service a structure is defined and can
be extendend in future.
Useless print (e.g., ceph orch status) are removed from
this role.

Signed-off-by: Francesco Pantano <fpantano@redhat.com>